+++ invoicedapi/Entities/Invoice.cs
@@ -97,7 +97,7 @@ namespace Invoiced
 		public string Notes { get; set; }
 
 		[JsonProperty("subtotal")]
-		public long? Subtotal { get; set; }
+		public double? Subtotal { get; set; }
 
 		[JsonProperty("discounts")]
 		public IList<Discount> Discounts { get; set; }
@@ -106,10 +106,10 @@ namespace Invoiced
 		public IList<Tax> Taxes { get; set; }
 
 		[JsonProperty("total")]
-		public long? Total { get; set; }
+		public double? Total { get; set; }
 
 		[JsonProperty("balance")]
-		public long? Balance { get; set; }
+		public double? Balance { get; set; }
 
 		[JsonProperty("payment_plan")]
 		public long? PaymentPlan { get; set; }
